Added master dimmer, manual ambient mode and blackout

+++ src/main/kotlin/de/randombyte/lighthub/Transition.kt
@@ -2,7 +2,7 @@ package de.randombyte.lighthub
 
 import de.randombyte.lighthub.osc.OscChannel
 
-abstract class Transition(val changeTo: Map<OscChannel, UByte>, val durationMs: Int) : Animation() {
+abstract class Transition(val changeTo: Map<OscChannel, Int>, val durationMs: Int) : Animation() {
     val firstValues = changeTo.keys.map { it.lastValue }
 
     abstract val finished: Boolean
